refactor(fetcher): extract CryptoFetcher and add _yf_last_prices()
Move all CoinGecko logic from fetcher.py into a new crypto_fetcher.py
module.  fetcher.py re-exports CryptoFetcher for backward compatibility
and gains the _yf_last_prices() helper that DRYs up the
normalize→download→last-close pipeline.
